+++ scripts/constants.js
@@ -1,3 +1,5 @@
+import actions from 'actions';
+
 // flows
 export const NUM_COLUMNS = 4;
 export const NUM_NODES_SUMMARY = 10;
@@ -86,3 +88,52 @@ export const CHORD_COLORS = ['#ea6869', '#34444c'];
 // map
 export const CARTO_BASE_URL = 'https://p2cs-sei.carto.com/api/v1/map/';
 export const CARTO_NAMED_MAPS_BASE_URL = 'https://p2cs-sei.carto.com/api/v1/map/named/';
+
+// GA
+export const GA_ACTION_WHITELIST = [
+  {
+    type: actions.UPDATE_NODE_SELECTION,
+    getPayload: action => action.data.map(d => d.name).join(',')
+  },
+  {
+    type: actions.SELECT_BIOME_FILTER,
+    getPayload: action => action.biomeFilter
+  },
+  {
+    type: actions.SELECT_YEARS,
+    getPayload: action => action.years.join(',')
+  },
+  {
+    type: actions.SELECT_RECOLOR_BY,
+    getPayload: action => action.value
+  },
+  {
+    type: actions.SELECT_QUANT,
+    getPayload: action => action.quant
+  },
+  {
+    type: actions.SELECT_VIEW,
+    getPayload: action => (action.detailedView) ? 'detailed' : 'overview'
+  },
+  {
+    type: actions.SELECT_COLUMN,
+    getPayload: (action, state) => {
+      return state.columns[action.columnId].name;
+    }
+  },
+  {
+    type: actions.TOGGLE_MAP
+  },
+  {
+    type: actions.TOGGLE_MAP_LAYERS_MENU
+  },
+  {
+    type: actions.SELECT_VECTOR_LAYERS,
+    getPayload: action => [action.layerData.title, action.layerData.direction].join(' - ')
+  },
+  {
+    type: actions.SELECT_CONTEXTUAL_LAYERS,
+    getPayload: action => action.contextualLayers.join(', ')
+  }
+
+];