+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.0.5
 - map: fixes a major issue where user was able to query a node without links by clicking a polygon on the map
+- sankey: Implemented Google Analytics events
 
 ### 1.0.4
 - sankey: Fixed issue with wrong merging of links coloured by node selection